ovos-voice-agent/AgentVoiceBoxEngine/backend/requirements.txt
# Django Core
Django>=5.1,<5.2
django-ninja>=1.3,<2.0
channels>=4.0,<5.0
channels-redis>=4.2,<5.0

# Database
psycopg[binary]>=3.1,<4.0
django-redis>=5.4,<6.0

# Authentication
PyJWT>=2.8,<3.0
cryptography>=41.0,<43.0

# HTTP Client
httpx>=0.27,<1.0

# Speech-to-Text
faster-whisper>=1.0,<2.0

# Validation
pydantic>=2.5,<3.0
pydantic-settings>=2.1,<3.0
email-validator>=2.1,<3.0

# Logging
django-structlog>=8.0,<9.0
structlog>=24.1,<25.0
python-json-logger>=2.0,<3.0

# Metrics
django-prometheus>=2.3,<3.0

# CORS
django-cors-headers>=4.3,<5.0

# Static Files
whitenoise>=6.6,<7.0

# External Integrations
hvac>=2.1,<3.0  # HashiCorp Vault
temporalio>=1.5,<2.0  # Temporal
confluent-kafka>=2.3,<3.0  # Kafka

# ASGI Server
uvicorn[standard]>=0.27,<1.0
gunicorn>=21.0,<23.0

# Development
ruff>=0.3,<1.0
black>=24.0,<25.0
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-asyncio>=0.23,<1.0
pytest-xdist>=3.5,<4.0  # Parallel test execution
hypothesis[django]>=6.98,<7.0  # Django-specific property testing with from_model()
